Added config to discover tests in any subdirectory
A common practice is to have unit tests sitting next to the thing it tests. For example, https://medium.com/@JeffLombardJr/organizing-tests-in-jest-17fc431ff850 

It took me a while to get this working. This PR adds a section to the mocha readme that allows such test discovery while ignoring the ones inside node_modules folders.