ng serve and background task
https://github.com/Microsoft/vscode/issues/64447

The tutorial as it stands creates a bit of ambiguity by defining a problem matcher for a background task, but not running the task in the background. Also, ng serve task was not defined as a preLaunch task. The guide instructed a user to follow a clunky flow.

Also, "identifiers" are deprecated and serve only to confuse new users. Therefore updated.